requirements.txt
asgiref==3.9.1
Django==5.2.6
django-parler==2.3
sqlparse==0.5.3
